Branch Main: add screenshot method to BaseTest
