+++ com.hillel/src/test/java/com/hillel/BaseTest.java
@@ -1,0 +1,52 @@
+package com.hillel;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import io.qameta.allure.Attachment;
+import lombok.SneakyThrows;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.ITestResult;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeClass;
+import pageObjects.MainPage;
+
+import java.io.File;
+import java.nio.file.Files;
+
+public class BaseTest {
+    protected WebDriver driver;
+
+    @BeforeClass
+    public void beforeClass() {
+        WebDriverManager.chromedriver().setup();
+        driver = new ChromeDriver();
+        driver.manage().window().maximize();
+    }
+
+    @AfterClass(alwaysRun = true)
+    public void afterClass() {
+        driver.quit();
+    }
+
+    @SneakyThrows
+    @AfterMethod
+    public void takeScreenshot(ITestResult result) {
+        if (!result.isSuccess()) {
+          File screenshot = ((TakesScreenshot)driver).getScreenshotAs(OutputType.FILE);
+          saveScreenshot(Files.readAllBytes(screenshot.toPath()));
+        }
+    }
+
+    public MainPage openApp() {
+        driver.get("https://the-internet.herokuapp.com/");
+        return new MainPage(driver);
+    }
+
+    @Attachment(value = "Page screenshot", type = "image/png")
+    public byte[] saveScreenshot(byte[] screenShot) {
+        return screenShot;
+    }
+}
