--- com.hillel/src/test/java/com/hillel/lesson18/BaseTest.java
@@ -1,29 +1,0 @@
-package com.hillel.lesson18;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import pageObjects.MainPage;
-
-public class BaseTest {
-    WebDriver driver;
-
-    @BeforeClass
-    public void beforeClass() {
-        WebDriverManager.chromedriver().setup();
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-    }
-
-    @AfterClass(alwaysRun = true)
-    public void afterClass() {
-        driver.quit();
-    }
-
-    public MainPage openApp() {
-        driver.get("https://the-internet.herokuapp.com/");
-        return new MainPage(driver);
-    }
-}
