+++ com.hillel/src/test/java/com/hillel/lesson18/TestCheckLogin.java
@@ -1,5 +1,6 @@
 package com.hillel.lesson18;
 
+import com.hillel.BaseTest;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import pageObjects.MainPage;
