+++ com.hillel/src/test/java/com/hillel/lesson18/TestTask1.java
@@ -1,5 +1,6 @@
 package com.hillel.lesson18;
 
+import com.hillel.BaseTest;
 import org.testng.annotations.Test;
 import pageObjects.ChallengingDomPage;
